+++ battlenet-api-browser/src/services/apiclientsettings.ts
@@ -1,3 +1,5 @@
 export class apiClientSettings {
     connectionType?: string;
+    disableSwitching: boolean = false;
+    disableSettings: boolean = false;
 }