+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added a simple query cache
 - npm run electron-make-deb now makes a .deb on compatible Linux systems (note - this is currently untested)
 - npm run electron-make-win now makes a Windows setup.exe (again, currently untested)
+- policy.jsonc can be used to override settings
 
 ##### Bug Fixes
 - Graceful error message if browser is unable to run Angular or doesn't have IndexedDB.